Move discovery of istiod dep into common method
