+++ business/mesh.go
@@ -211,46 +211,8 @@ func (in *MeshService) ResolveKialiControlPlaneCluster(r *http.Request) (*Cluste
 
 	// The "cluster_id" is set in an environment variable of
 	// the "istiod" deployment. Let's try to fetch it.
-	var istioDeployment *v1.Deployment
-	istioDeploymentConfig := conf.ExternalServices.Istio.IstiodDeploymentName
-	var err error
-
-	if IsNamespaceCached(conf.IstioNamespace) {
-		istioDeployment, err = kialiCache.GetDeployment(conf.IstioNamespace, istioDeploymentConfig)
-	} else {
-		istioDeployment, err = in.k8s.GetDeployment(conf.IstioNamespace, istioDeploymentConfig)
-	}
-	if err != nil && !errors.IsNotFound(err) {
-		return nil, err
-	}
-
-	if istioDeployment == nil {
-		kialiControlPlaneClusterCached = true
-		return nil, nil
-	}
-
-	if len(istioDeployment.Spec.Template.Spec.Containers) == 0 {
-		kialiControlPlaneClusterCached = true
-		return nil, nil
-	}
-
-	myClusterName := ""
-	for _, v := range istioDeployment.Spec.Template.Spec.Containers[0].Env {
-		if v.Name == "CLUSTER_ID" {
-			myClusterName = v.Value
-			break
-		}
-	}
-
-	gatewayToNamespace := false
-	for _, v := range istioDeployment.Spec.Template.Spec.Containers[0].Env {
-		if v.Name == "PILOT_SCOPE_GATEWAY_TO_NAMESPACE" {
-			gatewayToNamespace = v.Value == "true"
-			break
-		}
-	}
-
-	if len(myClusterName) == 0 {
+	myClusterName, gatewayToNamespace, err := kubernetes.ClusterInfoFromIstiod(*conf, in.k8s)
+	if err != nil {
 		// We didn't find it. This may mean that Istio is not setup with multi-cluster enabled.
 		kialiControlPlaneClusterCached = true
 		return nil, nil
