+++ kiali_test.go
@@ -5,14 +5,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-	apps_v1 "k8s.io/api/apps/v1"
-	core_v1 "k8s.io/api/core/v1"
-	meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-
 	"github.com/kiali/kiali/config"
-	"github.com/kiali/kiali/kubernetes/kubetest"
 	"github.com/kiali/kiali/util"
 )
 
@@ -94,70 +87,4 @@ func TestValidateAuthStrategy(t *testing.T) {
 			t.Errorf("Auth Strategy validation should have failed [%v]", conf.Auth.Strategy)
 		}
 	}
-}
-
-func TestGetClusterInfoFromIstiod(t *testing.T) {
-	require := require.New(t)
-	assert := assert.New(t)
-
-	conf := config.NewConfig()
-	k8s := kubetest.NewFakeK8sClient(
-		&core_v1.Namespace{ObjectMeta: meta_v1.ObjectMeta{Name: "istio-system"}},
-		&apps_v1.Deployment{
-			ObjectMeta: meta_v1.ObjectMeta{
-				Name:      "istiod",
-				Namespace: "istio-system",
-			},
-			Spec: apps_v1.DeploymentSpec{
-				Template: core_v1.PodTemplateSpec{
-					Spec: core_v1.PodSpec{
-						Containers: []core_v1.Container{
-							{
-								Name: "istiod",
-								Env: []core_v1.EnvVar{
-									{
-										Name:  "CLUSTER_ID",
-										Value: "east",
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-	)
-	clusterID, err := getClusterInfoFromIstiod(*conf, k8s)
-	require.NoError(err)
-
-	assert.Equal("east", clusterID)
-}
-
-func TestGetClusterInfoFromIstiodFails(t *testing.T) {
-	require := require.New(t)
-
-	conf := config.NewConfig()
-	k8s := kubetest.NewFakeK8sClient(
-		&core_v1.Namespace{ObjectMeta: meta_v1.ObjectMeta{Name: "istio-system"}},
-		&apps_v1.Deployment{
-			ObjectMeta: meta_v1.ObjectMeta{
-				Name:      "istiod",
-				Namespace: "istio-system",
-			},
-			Spec: apps_v1.DeploymentSpec{
-				Template: core_v1.PodTemplateSpec{
-					Spec: core_v1.PodSpec{
-						Containers: []core_v1.Container{
-							{
-								Name: "istiod",
-								Env:  []core_v1.EnvVar{},
-							},
-						},
-					},
-				},
-			},
-		},
-	)
-	_, err := getClusterInfoFromIstiod(*conf, k8s)
-	require.Error(err)
 }
