+++ kubernetes/cache/cache.go
@@ -86,13 +86,15 @@ func NewKialiCache(clientFactory kubernetes.ClientFactory, cfg config.Config, na
 	for cluster, client := range clientFactory.GetSAClients() {
 		cache, err := NewKubeCache(client, cfg, NewRegistryHandler(kialiCacheImpl.RefreshRegistryStatus), namespaceSeedList...)
 		if err != nil {
-			log.Errorf("[Kiali Cache] Error creating kube cache for cluster: %s. Err: %v", cluster, err)
+			log.Errorf("[Kiali Cache] Error creating kube cache for cluster: [%s]. Err: %v", cluster, err)
 			return nil, err
 		}
+		log.Infof("[Kiali Cache] Kube cache is active for cluster: [%s] and namespaces: %v", cluster, namespaceSeedList)
+
 		kialiCacheImpl.kubeCache[cluster] = cache
 
 		// TODO: Treat all clusters the same way.
-		if homeClient := clientFactory.GetSAHomeClusterClient(); homeClient != nil && homeClient.GetClusterInfo().Name == cluster {
+		if cluster == cfg.KubernetesConfig.ClusterName {
 			kialiCacheImpl.KubeCache = cache
 		}
 	}
